OFQuality
Floodlight controller module for the management of QoS in Software Defined-Networking.

Proxy for using W3C WebDriver-compatible clients to interact with Gecko-based browsers.

This program provides the HTTP API described by the WebDriver protocol to communicate with Gecko browsers, such as Firefox. It translates calls into the Firefox remote protocol by acting as a proxy between the local- and remote ends.

You can consult the change log for a record of all notable changes to the program. Releases are made available on GitHub on supported platforms.

The canonical source code repository for geckodriver now lives in mozilla-central under testing/geckodriver. You can read more about working with Mozilla source code on MDN. This means we do no longer accept pull requests on GitHub. Patches should be uploaded to a bug in the Testing :: GeckoDriver component.
Supported clients

Selenium users must update to version 3.5 or later to use geckodriver. Other clients that follow the W3C WebDriver specification are also supported.
Supported Firefoxen

geckodriver is not yet feature complete. This means that it does not yet offer full conformance with the WebDriver standard or complete compatibility with Selenium. You can track the implementation status of the latest Firefox Nightly on MDN. We also keep track of known Selenium, remote protocol, and specification problems in our issue tracker.

Support is best in Firefox 55 and greater, although generally the more recent the Firefox version, the better the experience as they have more bug fixes and features. Some features will only be available in the most recent Firefox versions, and we strongly advise using the latest Firefox Nightly with geckodriver. Since Windows XP support in Firefox was dropped with Firefox 53, we do not support this platform.